Class mode and AI feature
