+++ Bingo/Application.cpp
@@ -1,0 +1,38 @@
+#include "Game.h"
+int main()
+{
+    Game game;
+    game.start();
+
+    //player userA, userB;
+
+    //cout << "Player A........\n";
+    //PlayerInfo(userA);
+    //clearScreen(); // <--- here
+    //clearScreen(); // <--- here
+    //cout << "Player B........\n";
+    //PlayerInfo(userB);
+
+    //players.push_back(&userA);
+
+    //players.push_back(&userB);
+
+    //player* user = &userA;
+
+    //while (!gameComplete(userA, userB))
+    //{
+    //    clearScreen(); // <--- here
+    //    cout << "-------------------------\n";
+    //    cout << "Player " << user->playerName << "'s turn\n";
+    //    chance(user);
+    //    user = (user == &userA) ? &userB : &userA;
+    //}
+
+    //cout << "\n Game Over!\n";
+    //if (checkRow(userA) + checkCol(userA) + diagCount(userA) >= 5)
+    //    cout << userA.playerName << " wins!\n";
+    //else
+    //    cout << userB.playerName << " wins!\n";
+
+    //return 0;
+}
