+++ Bingo/utils.cpp
@@ -1,0 +1,21 @@
+#include "utils.h"
+#include <cstdlib>
+#include <limits>
+
+void clearScreen() {
+#ifdef _WIN32
+    system("cls");
+#else
+    system("clear");
+#endif
+}
+
+int safeInput() {
+    int val;
+    while (!(cin >> val)) {
+        cout << "Invalid input. Try again: ";
+        cin.clear();
+        cin.ignore(numeric_limits<streamsize>::max(), '\n');
+    }
+    return val;
+}
